#111 refactor: Move files to subfolders and extract api request as well as the interfaces
