+++ src/features/course-admin/module/ModuleDetail.tsx
