+++ src/features/course-admin/module/ModuleList.module.scss
@@ -41,12 +41,13 @@
     color: red;
     background: none;
     border: none;
-    padding: 0;
     font: inherit;
     cursor: pointer;
     outline: inherit;
+    box-shadow: none;
+    padding-left: 0;
 }
 
 .alignSubmenu {
-    text-align: left;
+    text-align: center;
 }
